requirements.txt
numpy>=1.24.0
cupy-cuda12x>=12.0.0
numba>=0.58.0
pygame>=2.5.0
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
scipy>=1.11.0 